Created foundations to define dynamic routes
