+++ src/pages/documentation/classic-web-server.tsx
@@ -1,7 +1,8 @@
 import React, { FunctionComponent } from 'react';
 import ClassicWebServerContent from '../../content/documentation/classic-web-server.mdx';
 import Breadcrumb from '../../components/Breadcrumb';
-import Documentation, { Item } from './index';
+import { Item } from '../../utils/docUtils';
+import Documentation from './index';
 
 const breadcrumbItems: Item[] = [{ text: 'Getting started', link: '/documentation' }];
