+++ src/pages/documentation/install-docker-image.tsx
@@ -1,7 +1,8 @@
 import React, { FunctionComponent } from 'react';
 import InstallDockerImageContent from '../../content/documentation/install-docker-image.mdx';
 import Breadcrumb from '../../components/Breadcrumb';
-import Documentation, { Item } from './index';
+import { Item } from '../../utils/docUtils';
+import Documentation from './index';
 
 const breadcrumbItems: Item[] = [{ text: 'Getting started', link: '/documentation' }];
