+++ src/pages/documentation/integrations/short-menu-ios.tsx
@@ -1,7 +1,8 @@
 import React, { FunctionComponent } from 'react';
 import ShortMenuIosContent from '../../../content/integrations/short-menu-ios.mdx';
 import Breadcrumb from '../../../components/Breadcrumb';
-import Documentation, { Item } from '../index';
+import { Item } from '../../../utils/docUtils';
+import Documentation from '../index';
 
 const breadcrumbItems: Item[] = [
   { text: 'Docs', link: '/documentation' },
