+++ src/pages/documentation/integrations/short-menu-mac.tsx
@@ -1,7 +1,8 @@
 import React, { FunctionComponent } from 'react';
 import ShortMenuMacContent from '../../../content/integrations/short-menu-mac.mdx';
 import Breadcrumb from '../../../components/Breadcrumb';
-import Documentation, { Item } from '../index';
+import { Item } from '../../../utils/docUtils';
+import Documentation from '../index';
 
 const breadcrumbItems: Item[] = [
   { text: 'Docs', link: '/documentation' },
