+++ src/pages/documentation/shlink-web-client/index.tsx
@@ -1,7 +1,8 @@
 import React, { FunctionComponent } from 'react';
 import ShlinkWebClientContent from '../../../content/shlink-web-client/index.mdx';
 import Breadcrumb from '../../../components/Breadcrumb';
-import Documentation, { Item } from '../index';
+import { Item } from '../../../utils/docUtils';
+import Documentation from '../index';
 
 const breadcrumbItems: Item[] = [{ text: 'Docs', link: '/documentation' }];
