+++ src/pages/documentation/shlink-web-client/serve-in-sub-path.tsx
@@ -1,6 +1,7 @@
 import React, { FunctionComponent } from 'react';
 import ServeInSubPathContent from '../../../content/shlink-web-client/serve-in-sub-path.mdx';
-import Documentation, { Item } from '../index';
+import { Item } from '../../../utils/docUtils';
+import Documentation from '../index';
 import Breadcrumb from '../../../components/Breadcrumb';
 
 const breadcrumbItems: Item[] = [
